\name{Anthropometry-internalPlotTree}
\alias{make.circle.discovery}
\alias{make.arrow.circle}
\title{
Several internal functions used to build the HIPAM plot tree
}
\description{
This file contains two subfunctions called by the \code{\link{plotTreeHipamAnthropom}} function. Therefore, they are not used directly. These functions are the same as those called by the \emph{plot.tree} function of the \pkg{smida} R package (Wit et al. (2006)).
}
\references{
Vinue, G., Leon, T., Alemany, S., and Ayala, G., (2013). Looking for representative fit models for apparel sizing, \emph{Decision Support Systems} \bold{57}, 22--33.

Wit, E., and McClure, J., (2004). \emph{Statistics for Microarrays: Design, Analysis and Inference}. John Wiley & Sons, Ltd.

Wit, E., and McClure, J., (2006). Statistics for Microarrays: Inference, Design and Analysis. R package version 0.1. \url{http://www.math.rug.nl/~ernst/book/smida.html}.
}
\author{
All these functions were originally created by E. Wit et al., and they are available freely on
\url{http://www.math.rug.nl/~ernst/book/smida.html}.
}
\seealso{
\code{\link{plotTreeHipamAnthropom}}, \code{\link{hipamAnthropom}}
}
\keyword{internal}

