\name{cube34}
\alias{cube34}
\docType{data}
\title{
Cube of 34 landmarks
}
\description{
This is a cube made up of 34 landmarks, used as controlled data in the simulation study carried out in the paper referred below.
}
\usage{cube34}
\format{An array with one matrix of 34 rows and 3 columns.}
\source{Software Rhinoceros.}
\references{
Vinue, G., Simo, A., and Alemany, S., (2014). The k-means algorithm for 3D shapes with an application to apparel design. Accepted for publication in Advances in Data Analysis and Classification.
}
\keyword{datasets}
