\name{cube8landm}
\alias{cube8landm}
\docType{data}
\title{
Cube of 8 landmarks
}
\description{
This is a cube made up of 8 landmarks, used as controlled data in the simulation study carried out in the paper referred below.
}
\usage{cube8landm}
\format{An array with one matrix of 8 rows and 3 columns.}
\source{Software Rhinoceros.}
\references{
Vinue, G., Simo, A., and Alemany, S., (2016). The k-means algorithm for 3D shapes with an application to apparel design, \emph{Advances in Data Analysis and Classification} \bold{10(1)}, 103--132.
}
\keyword{datasets}
