\name{dataDemo}
\alias{dataDemo}
\docType{data}
\title{
Demo database of the Spanish anthropometric survey
}
\description{
This a demo database for academic and training purposes. It is oriented to exemplify the use of \code{\link{trimowa}}, \code{\link{hipamAnthropom}} and \code{\link{TDDclust}}. 

It is made up of 600 women selected randomly from the Spanish anthropometric survey and five anthropometric variables: chest circumference, neck to ground length, waist circumference, hip circumference and bust circumference. These variables have been chosen following the recommendations of experts. In addition, they are commonly used in the literature about sizing system design and they appear in the European standard to sizing system.
}
\usage{dataDemo}
\format{A matrix with 600 rows and 5 columns. Each row corresponds to an observation, and each column corresponds to a variable.}
\source{Anthropometric survey of the Spanish female population.}
\references{
Alemany, S., Gonzalez, J. C., Nacher, B., Soriano, C., Arnaiz, C., and Heras, H., (2010). Anthropometric survey of the Spanish female population aimed at the apparel industry. \emph{Proceedings of the 2010 Intl. Conference on 3D Body scanning Technologies}, 307--315.

Ibanez, M. V., Vinue, G., Alemany, S., Simo, A., Epifanio, I., Domingo, J., and Ayala, G., (2012). Apparel sizing using trimmed PAM and OWA operators, \emph{Expert Systems with Applications} \bold{39}, 10512--10520.

Vinue, G., Leon, T., Alemany, S., and Ayala, G., (2013). Looking for representative fit models for apparel
sizing, \emph{Decision Support Systems} \bold{57}, 22--33.

Vinue, G., and Ibanez, M. V., (2014). \emph{Data depth and Biclustering applied to anthropometric data. Exploring their utility in apparel design}. In progress.

European Committee for Standardization. Size designation of clothes. Part 2: Primary and secondary dimensions. (2002).

European Committee for Standardization. Size designation of clothes. Part 3: Measurements and intervals. (2005).
}
\seealso{
\code{\link{trimowa}}, \code{\link{hipamAnthropom}}, \code{\link{TDDclust}}
}
\keyword{datasets}
