\name{parallelep34landm}
\alias{parallelep34landm}
\docType{data}
\title{
Parallelepiped of 34 landmarks
}
\description{
This is a parallelepiped made up of 34 landmarks, used as controlled data in the simulation study carried out in the paper referred below.
}
\usage{parallelep34landm}
\format{An array with one matrix of 34 rows and 3 columns.}
\source{Software Rhinoceros.}
\references{
Vinue, G., Simo, A., and Alemany, S., (2016). The k-means algorithm for 3D shapes with an application to apparel design, \emph{Advances in Data Analysis and Classification} \bold{10(1)}, 103--132.
}
\keyword{datasets}
