\name{parallelep8landm}
\alias{parallelep8landm}
\docType{data}
\title{
Parallelepiped of 8 landmarks
}
\description{
This is a parallelepiped made up of 8 landmarks, used as controlled data in the simulation study carried out in the paper referred below.
}
\usage{parallelep8landm}
\format{An array with one matrix of 8 rows and 3 columns.}
\source{Software Rhinoceros.}
\references{
Vinue, G., Simo, A., and Alemany, S., (2014). The k-means algorithm for 3D shapes with an application to apparel design, \emph{Advances in Data Analysis and Classification}, 1--30.
}
\keyword{datasets}
