\name{parallelepiped34}
\alias{parallelepiped34}
\docType{data}
\title{
Parallelepiped of 34 landmarks
}
\description{
This is a parallelepiped made up of 34 landmarks, used as controlled data in the simulation study carried out in the paper referred below.
}
\usage{parallelepiped34}
\format{An array with one matrix of 34 rows and 3 columns.}
\source{Software Rhinoceros.}
\references{
Vinue, G., Simo, A., and Alemany, S., (2014). The k-means algorithm for 3D shapes with an application to apparel design. Submitted for publication.
}
\keyword{datasets}
