\name{USAFSurvey}
\alias{USAFSurvey}
\docType{data}
\title{
USAF 1967 survey 
}
\description{
This data set comes from the 1967 United States Air Force (USAF) survey (available from \url{http://www.dtic.mil/dtic/}). The 1967 USAF survey was conducted during the first three months of 1967 under the direction of the Anthropology Branch of the Aerospace Medical Research Laboratory, located in Ohio. Subjects were measured at 17 Air Force bases across the United States of America. A total of 202 variables (including body dimensions and background variables) were taken on 2420 Air Force personnel between 21 and 50 years of age. 

Please find in www.uv.es/vivigui/softw/data_information.zip some files that provide a detailed information about this database.

In Epifanio (2013), the column numbers selected were c(48,40,39,33,32) and correspond to 'Thumb tip reach', 'Buttock-Knee length', 'Popliteal height sitting', 'Sitting height', 'Eye height sitting' and 'Shoulder height sitting'.
}
\usage{USAFSurvey}
\format{A matrix with 2420 rows and 202 columns. Each row corresponds to an observation, and each column corresponds to a variable.}
\source{1967 United States Air Force (USAF) survey.}
\references{
Vinue, G., Epifanio, I., and Alemany, S., (2015). Archetypoids: a new approach to define representative archetypal data, \emph{Computational Statistics and Data Analysis} \bold{87}, 102--115.

Epifanio, I., Vinue, G., and Alemany, S., (2013). Archetypal analysis: contributions for estimating boundary cases in multivariate accommodation problem, \emph{Computers & Industrial Engineering} \bold{64}, 757--765.
}
\keyword{datasets}
